added ngfb_preserve option
git-svn-id: https://plugins.svn.wordpress.org/nextgen-facebook/trunk@720490 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ nextgen-facebook.php
@@ -126,7 +126,8 @@ if ( ! class_exists( 'ngfbPlugin' ) ) {
 
 		// delete options table entries only when plugin deactivated and deleted
 		public function uninstall() {
-			delete_option( NGFB_OPTIONS_NAME );
+			if ( empty( $this->options['ngfb_preserve'] ) )
+				delete_option( NGFB_OPTIONS_NAME );
 		}
 
 		// called by WP init action
